+++ concrete_datastore/api/v1/views.py
@@ -77,6 +77,7 @@ from concrete_datastore.api.v1.filters import (
     FilterSupportingComparaisonBackend,
     FilterForeignKeyIsNullBackend,
     FilterSupportingForeignKey,
+    FilterSupportingManyToMany,
 )
 from concrete_datastore.api.v1.authentication import (
     TokenExpiryAuthentication,
@@ -1225,6 +1226,7 @@ class PaginatedViewSet(object):
         FilterSupportingComparaisonBackend,
         FilterForeignKeyIsNullBackend,
         FilterSupportingForeignKey,
+        FilterSupportingManyToMany,
     )
     filterset_fields = ()
     ordering_fields = '__all__'