+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Added
 
-- nothing added
+- Add API filters for ManyToMany fields using the `__in` filter
 
 ### Changed